renames package resource => resources
