+++ src/dimcat/data/catalogs/base.py
@@ -7,8 +7,8 @@ import frictionless as fl
 import marshmallow as mm
 from dimcat.data.base import Data
 from dimcat.data.packages.base import Package, PackageSpecs
-from dimcat.data.resource.base import Resource
-from dimcat.data.resource.features import FeatureSpecs
+from dimcat.data.resources.base import Resource
+from dimcat.data.resources.features import FeatureSpecs
 from dimcat.dc_exceptions import (
     DuplicatePackageNameError,
     EmptyCatalogError,
