+++ src/dimcat/data/catalogs/inputs.py
@@ -1,8 +1,8 @@
 from __future__ import annotations
 
 from dimcat.data.catalogs.base import DimcatCatalog
-from dimcat.data.resource.base import SomeDataframe
-from dimcat.data.resource.features import Feature, FeatureSpecs, feature_specs2config
+from dimcat.data.resources.base import SomeDataframe
+from dimcat.data.resources.features import Feature, FeatureSpecs, feature_specs2config
 
 
 class InputsCatalog(DimcatCatalog):
