+++ src/dimcat/data/catalogs/outputs.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from typing import Iterator, Tuple
 
 from dimcat.data.catalogs.base import DimcatCatalog
-from dimcat.data.resource.dc import DimcatResource
-from dimcat.data.resource.features import FeatureSpecs, feature_specs2config
+from dimcat.data.resources.dc import DimcatResource
+from dimcat.data.resources.features import FeatureSpecs, feature_specs2config
 
 
 class OutputsCatalog(DimcatCatalog):
