+++ src/dimcat/data/resources/base.py
