+++ src/dimcat/data/resources/features.py
@@ -7,7 +7,7 @@ from typing import Iterable, List, MutableMapping, Optional, TypeAlias, Union
 import frictionless as fl
 import marshmallow as mm
 from dimcat.base import DimcatConfig, ObjectEnum, is_subclass_of
-from dimcat.data.resource.dc import DimcatResource
+from dimcat.data.resources.dc import DimcatResource
 from dimcat.dc_exceptions import ResourceNotProcessableError
 
 logger = logging.getLogger(__name__)
