+++ src/dimcat/data/resources/utils.py
