+++ src/dimcat/steps/analyzers/base.py
@@ -8,10 +8,10 @@ from typing import Any, ClassVar, Iterable, Optional, Type, TypeVar
 import marshmallow as mm
 from dimcat.base import ObjectEnum
 from dimcat.data.datasets.processed import AnalyzedDataset
-from dimcat.data.resource.base import SomeSeries
-from dimcat.data.resource.dc import DimcatResource
-from dimcat.data.resource.features import Feature, FeatureSpecs
-from dimcat.data.resource.results import Result
+from dimcat.data.resources.base import SomeSeries
+from dimcat.data.resources.dc import DimcatResource
+from dimcat.data.resources.features import Feature, FeatureSpecs
+from dimcat.data.resources.results import Result
 from dimcat.steps.base import FeatureProcessingStep
 
 logger = logging.getLogger(__name__)
