+++ src/dimcat/steps/analyzers/counters.py
@@ -1,8 +1,8 @@
 import logging
 
-from dimcat.data.resource.base import SomeDataframe, SomeSeries
-from dimcat.data.resource.dc import DimcatResource
-from dimcat.data.resource.features import Feature
+from dimcat.data.resources.base import SomeDataframe, SomeSeries
+from dimcat.data.resources.dc import DimcatResource
+from dimcat.data.resources.features import Feature
 from dimcat.steps.analyzers.base import Analyzer
 
 logger = logging.getLogger(__name__)
