+++ src/dimcat/steps/base.py
@@ -18,14 +18,14 @@ from dimcat.base import DimcatConfig, DimcatObject, get_class
 from dimcat.data.base import Data
 from dimcat.data.datasets.base import Dataset
 from dimcat.data.packages.dc import DimcatPackage
-from dimcat.data.resource.base import Resource, ResourceSpecs, resource_specs2resource
-from dimcat.data.resource.dc import DimcatResource
-from dimcat.data.resource.features import (
+from dimcat.data.resources.base import Resource, ResourceSpecs, resource_specs2resource
+from dimcat.data.resources.dc import DimcatResource
+from dimcat.data.resources.features import (
     FeatureName,
     FeatureSpecs,
     features_argument2config_list,
 )
-from dimcat.data.resource.utils import ensure_level_named_piece
+from dimcat.data.resources.utils import ensure_level_named_piece
 from dimcat.dc_exceptions import (
     EmptyDatasetError,
     EmptyResourceError,
