+++ src/dimcat/steps/extractors/base.py
@@ -4,7 +4,7 @@ from typing import Iterable
 
 from dimcat.base import DimcatConfig
 from dimcat.data.datasets.base import Dataset
-from dimcat.data.resource.dc import DimcatResource
+from dimcat.data.resources.dc import DimcatResource
 from dimcat.steps.base import FeatureProcessingStep
 from marshmallow import fields, validate
