+++ tests/test_pipeline.py
@@ -1,5 +1,5 @@
 from dimcat.base import DimcatConfig
-from dimcat.data.resource.dc import PieceIndex
+from dimcat.data.resources.dc import PieceIndex
 from dimcat.steps.base import FeatureProcessingStep
 from dimcat.steps.extractors.base import FeatureExtractor
 from dimcat.steps.groupers.base import CustomPieceGrouper
